// src/types/enums/product.enum.ts
export enum PaymentMethodsEnum {
  CASH = 'cash',
  BANK_TRANSFER = 'bank_transfer',
  VF_CASH = 'vf_cash',
}

export enum InstallmentTypeEnum {
  WEEKLY = 'weekly',
  MONTH = 'month',
  QUARTER = 'quarter',
  HALFYEAR = 'halfyear',
  YEAR = 'year',
}

export enum PaidStatusEnum {
  INSTALLMENTS = 'installments',
  PAID = 'paid',
  PENDING = 'pending',
}

export enum SizesEnum {
  XS = 'XS',
  S = 'S',
  M = 'M',
  L = 'L',
  XL = 'XL',
  XXL = 'XXL',
  XXXL = 'XXXL',
}

export enum PeriodsEnum {
  MONTH = 'month',
  QUARTER = 'quarter',
  HALFYEAR = 'halfyear',
  YEAR = 'year',
}

export enum CarType {
  SEDAN = 'Sedan',
  HATCHBACK = 'Hatchback',
  SUV = 'SUV',
  CROSSOVER = 'Crossover',
  COUPE = 'Coupe',
  CONVERTIBLE = 'Convertible',
  PICKUP = 'Pickup',
  MINIVAN = 'Minivan',
  VAN = 'Van',
  STATION_WAGON = 'StationWagon',
  TRUCK = 'Truck',
  OFFROAD = 'OffRoad',
  ELECTRIC = 'Electric',
}